requirements.txt
django-debug-toolbar
mysql-python
django-tastypie
# django >= 1.8.8